drakrun: attach snapshot modification time to analysis reports
